requirements.txt
flask==3.1.0
PyMuPDF==1.25.3
python-pptx==1.0.2
Pillow==11.1.0
google-generativeai==0.8.4
python-dotenv==1.0.1
